NF: finishedMsg in Abstract
