+++ AnkiDroid/src/main/java/com/ichi2/libanki/sched/Sched.java
@@ -1198,18 +1198,6 @@ public class Sched extends SchedV2 {
      * *****************************************
      */
 
-    @Override
-    public CharSequence finishedMsg(Context context) {
-        SpannableStringBuilder sb = new SpannableStringBuilder();
-        sb.append(context.getString(R.string.studyoptions_congrats_finished));
-        StyleSpan boldSpan = new StyleSpan(Typeface.BOLD);
-        sb.setSpan(boldSpan, 0, sb.length(), 0);
-        sb.append(_nextDueMsg(context));
-        // sb.append("\n\n");
-        // sb.append(_tomorrowDueMsg(context));
-        return sb;
-    }
-
 
     @Override
     public boolean haveBuried() {
